pyproject.toml
[project]
name = "twilioconversationrelay"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "openai",
    "python-dotenv",
    "uvicorn[standard]",
]
